ibeeam/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
celery==5.2.7
django-environ==0.9.0
django-debug-toolbar==3.7.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.24.2
django-celery-results==2.4.0
django-mptt==0.14.0
psycopg2-binary==2.9.5
requests==2.28.1
Pillow==9.3.0
gunicorn==20.1.0
redis==4.3.5
python-jose==3.3.0
